+++ drivers/platform/Kconfig
@@ -16,6 +16,7 @@ source "drivers/platform/maxim/Kconfig"
 source "drivers/platform/pico/Kconfig"
 source "drivers/platform/aducm3029/Kconfig"
 source "drivers/platform/xilinx/Kconfig"
+source "drivers/platform/linux-userspace/Kconfig"
 
 config PLATFORM_I2C_BITBANG
 	default n
