+++ drivers/platform/linux-userspace/linux_timer.c
